+++ CHANGELOG.md
@@ -40,6 +40,50 @@ details.
   request and response provided at invocation with the
   `Zend\Stratigility\Http\*` variants, as these have been removed.
 
+- [#76](https://github.com/zendframework/zend-stratigility/pull/76) updates
+  `MiddlewarePipe` to implement only the http-interop
+  `ServerMiddlewareInterface`, and not the Stratigility-specific
+  `MiddlewareInterface` (which was removed).
+
+- [#76](https://github.com/zendframework/zend-stratigility/pull/76) updates
+  `Zend\Stratigility\Middleware\ErrorHandler` to implement the http-interop
+  `ServerMiddlewareInterface` instead of the Stratigility-specific
+  `MiddlewareInterface` (which was removed).
+
+- [#76](https://github.com/zendframework/zend-stratigility/pull/76) updates
+  `Zend\Stratigility\Middleware\NotFoundHandler` to implement the http-interop
+  `ServerMiddlewareInterface` instead of the Stratigility-specific
+  `MiddlewareInterface` (which was removed).
+
+- [#76](https://github.com/zendframework/zend-stratigility/pull/76) updates
+  `MiddlewarePipe::__invoke()` to require a third argument, now named
+  `$delegate`, and no longer type-hinted. If a callable not implementing
+  http-interop `DelegateInterface` is provided, it is wrapped in the
+  `CallableDelegateDecorator` (introduced in 1.3.0). The method then calls its
+  own `process()` method with the request and delegate. This method should
+  typically only be used as an entry point for an application.
+
+- [#76](https://github.com/zendframework/zend-stratigility/pull/76) updates
+  `MiddlewarePipe::pipe()` to raise an exception if callable middleware using
+  the legacy double-pass signature is provided, but no response prototype is
+  composed in the `MiddlewarePipe` instance yet.
+
+- [#76](https://github.com/zendframework/zend-stratigility/pull/76) updates
+  the constructor of `Next` to rename the `$done` argument to `$nextDelegate`
+  and typehint it against the http-interop `DelegateInterface`.
+
+- [#76](https://github.com/zendframework/zend-stratigility/pull/76) updates
+  `Next::__invoke()` to remove all arguments except the `$request` argument; the
+  method now proxies to the instance `process()` method.
+
+- [#76](https://github.com/zendframework/zend-stratigility/pull/76) updates
+  `Next` to no longer compose a `Dispatch` instance; it is now capable of
+  dispatching on its own.
+
+- [#76](https://github.com/zendframework/zend-stratigility/pull/76) updates the
+  `Zend\Stratigility\Route` constructor to raise an exception if
+  non-http-interop middleware is provided as the route handler.
+
 ### Deprecated
 
 - Nothing.
@@ -50,28 +94,46 @@ details.
   `Zend\Stratigility\FinalHandler`. Use `Zend\Stratigility\NoopFinalHandler`
   instead, along with `Zend\Stratigility\Middleware\ErrorHandler` and
   `Zend\Stratigility\Middleware\NotFoundHandler` (or equivalents).
+
 - [#67](https://github.com/zendframework/zend-stratigility/pull/67) removes
   `Zend\Stratigility\ErrorMiddlewareInterface`. Register middleware, such as
   `Zend\Stratigility\Middleware\ErrorHandler`, in outer layers of your
   application to handle errors.
+
 - [#67](https://github.com/zendframework/zend-stratigility/pull/67) removes
   `Zend\Stratigility\Dispatch`. This was an internal detail of the `Next`
   implementation, and should not affect most applications.
+
 - [#67](https://github.com/zendframework/zend-stratigility/pull/67) removes
   `Zend\Stratigility\Utils::getArity()`. This was used only in `Dispatch`;
   since middleware signatures no longer vary, it is no longer necessary.
+
 - [#67](https://github.com/zendframework/zend-stratigility/pull/67) removes
   the final, optional `$err` argument to `Zend\Stratigility\Next()`; raise
   exceptions instead, and provide error handling middleware such as
   `Zend\Stratigility\Middleware\ErrorHandler` instead.
+
 - [#67](https://github.com/zendframework/zend-stratigility/pull/67) removes
   the `$done` argument to the `Zend\Stratigility\Next` constructor.
+
 - [#71](https://github.com/zendframework/zend-stratigility/pull/71) removes
   the `Zend\Stratigility\Http\Request` class.
+
 - [#71](https://github.com/zendframework/zend-stratigility/pull/71) removes
   the `Zend\Stratigility\Http\Response` class.
+
 - [#71](https://github.com/zendframework/zend-stratigility/pull/71) removes
   `Zend\Stratigility\Http\ResponseInterface`.
+
+- [#76](https://github.com/zendframework/zend-stratigility/pull/76) removes
+  `Zend\Stratigility\MiddlewareInterface` and `Zend\Stratigility\ErrorMiddlewareInterface`.
+  The latter is removed entirely, while the former is essentially replaced by
+  http-interop's `ServerMiddlewareInterface`. You may still write callable
+  middleware using the legacy double-pass signature, however.
+
+- [#76](https://github.com/zendframework/zend-stratigility/pull/76) removes the
+  `Zend\Stratigility\Dispatch` class. The class was an internal detail of
+  `Next`, and no longer required.
 
 ### Fixed
 
@@ -120,6 +182,46 @@ details.
   - `originalResponse`, representing the response provided to it.
   - `originalUri`, representing URI instance composed in the request provided to it.
 
+- [#75](https://github.com/zendframework/zend-stratigility/pull/75) adds support
+  for [http-interop middleware 0.2.0](https://github.com/http-interop/http-middleware/tree/ff545c87e97bf4d88f0cb7eb3e89f99aaa53d7a9).
+  For full details, see the [migration guide](https://docs.zendframework.com/zend-stratigility/migration/to-v2/#http-interop-compatibility).
+  As a summary of features:
+  - You may now pipe http-interop middleware to `MiddlewarePipe` instances.
+  - You may now pipe callable middleware that defines the same signature as
+    http-interop middleware to `MiddlewarePipe` instances; these will be
+    decorated in a `Zend\Stratigility\Middleware\CallableInteropMiddlewareWrapper`
+    instance.
+  - `MiddlewarePipe` now implements the http-interop
+    `ServerMiddlewareInterface`, allowing it to be used in http-interop
+    middleware dispatchers.
+
+- [#75](https://github.com/zendframework/zend-stratigility/pull/75) adds the
+  class `Zend\Stratigility\Middleware\CallableMiddlewareWrapper`. It accepts
+  callable double-pass middleware and a response prototype, and implements the
+  http-interop `ServerMiddlewareInterface`, allowing you to adapt existing
+  callable middleware to work with http-interop middleware dispatchers.
+
+- [#75](https://github.com/zendframework/zend-stratigility/pull/75) adds the
+  class `Zend\Stratigility\Middleware\CallableInteropMiddlewareWrapper`. It accepts
+  callable middleware that follows the http-interop `ServerMiddlewareInterface`,
+  and implements that interface itself, to allow composing such middleware in
+  http-interop middleware dispatchers.
+
+- [#75](https://github.com/zendframework/zend-stratigility/pull/75) adds the
+  class `Zend\Stratigility\Delegate\CallableDelegateDecorator`, which can be
+  used to add http-interop middleware support to your existing callable
+  middleware.
+
+- [#75](https://github.com/zendframework/zend-stratigility/pull/75) adds a new
+  method to `MiddlewarePipe`, `setResponseProtoype()`. When this method is
+  invoked with a PSR-7 response, the following occurs:
+  - That response is injected in `Next` and `Dispatch` instances, to allow
+    dispatching legacy callable middleware as if it were http-interop
+    middleware.
+  - Any callable middleware implementing the legacy signature will now be
+    decorated using the above `CallableMiddlewareWrapper` in order to adapt it
+    as http-interop middleware.
+
 ### Changed
 
 - [#70](https://github.com/zendframework/zend-stratigility/pull/70) makes the
@@ -131,6 +233,26 @@ details.
   - It no longer writes to the response using the
     `Zend\Stratigility\Http\Response`-specific `write()` method, but rather
     pulls the message body and writes to that.
+
+- [#75](https://github.com/zendframework/zend-stratigility/pull/75) updates
+  `MiddlewarePipe` to inject the `$response` argument to `__invoke()` as the
+  response prototype.
+
+- [#75](https://github.com/zendframework/zend-stratigility/pull/75) updates
+  `Zend\Stratigility\Next` to implement the http-interop middleware
+  `DelegateInterface`. It also updates `Zend\Stratigility\Dispatch` to add a new
+  method, `process()`, following the `DelegateInterface` signature, thus
+  allowing `Next` to properly process http-interop middleware. These methods
+  will use the composed response prototype, if present, to invoke callable
+  middleware using the legacy signature.
+
+- [#75](https://github.com/zendframework/zend-stratigility/pull/75) updates
+  `Next` to allow the `$done` constructor argument to be an http-interop
+  `DelegateInterface`, and will invoke it as such if the queue is exhausted.
+
+- [#75](https://github.com/zendframework/zend-stratigility/pull/75) updates
+  `Route` (which is used internally by `MiddlewarePipe` to allow either callable
+  or http-interop middleware as route handlers.
 
 ### Deprecated
 
@@ -173,6 +295,17 @@ details.
   `Zend\Stratigility\Http\Response`. Additionally, the methods `write()`,
   `end()`, `isComplete()`, and `getOriginalResponse()` now emit deprecation
   notices when invoked, urging users to update their code.
+
+- [#75](https://github.com/zendframework/zend-stratigility/pull/75) deprecates
+  the `$response` argument in existing callable middleware. Please only operate
+  on the response returned by `$next`/`$delegate`, or create a response. See the
+  documentation [section on response arguments](https://docs.zendframework.com/zend-stratigility/api/#response-argument)
+  for more details.
+
+- [#75](https://github.com/zendframework/zend-stratigility/pull/75) deprecates
+  usage of error middleware, and thus deprecates the `$err` argument to `$next`;
+  explicitly invoking error middleware using that argument to `$next` will now
+  raise a deprecation notice.
 
 ### Removed